refactor: move to schema folder
